Updated the superuser container to accept the component and updated other files to reflect the change
