+++ src/Components/superuser/editContest/EditContestContainer.jsx
@@ -1,0 +1,38 @@
+import React from 'react';
+import { useQuery } from '@apollo/react-hooks';
+import { useParams } from 'react-router-dom';
+import { GET_CONTEST_EDIT_DETAILS } from '../../../graphql/queries';
+import SomethingWentWrong from '../../common/SomethingWentWrong/index';
+import useSessionExpired from '../../../customHooks/useSessionExpired';
+import Spinner from '../../common/Spinner/index';
+import EditContest from './EditContest';
+
+
+const EditcontestContainer = () => {
+  const { redirectOnSessionExpiredBeforeRender, isSessionExpired } = useSessionExpired();
+  const { contestId } = useParams();
+  const {
+    loading, error, data,
+  } = useQuery(GET_CONTEST_EDIT_DETAILS, {
+    variables: { code: contestId },
+  });
+
+  if (loading) return <Spinner />;
+  if (error) return <SomethingWentWrong message="An error has been encountered." />;
+  if (data.contestCode) {
+    const contestData = data.contestCode;
+    const { users } = data;
+    return (
+      <EditContest contestData={contestData} users={users} />
+    );
+  }
+  if (isSessionExpired(data.contestCode)) {
+    // since the component hasn't rendered or returned anything,
+    // we use redirectOnSessionExpiredBeforeRender function
+    return redirectOnSessionExpiredBeforeRender();
+  }
+  // case for the user not being admin or superuser
+  return <SomethingWentWrong message="An unexpected error has been encountered" />;
+};
+
+export default EditcontestContainer;
