+++ src/Components/superuser/editContest/index.jsx
@@ -1,41 +1,14 @@
 import React from 'react';
-import { useQuery } from '@apollo/react-hooks';
-import { useParams } from 'react-router-dom';
-import { GET_CONTEST_EDIT_DETAILS } from '../../../graphql/queries';
-import SomethingWentWrong from '../../common/SomethingWentWrong/index';
-import useSessionExpired from '../../../customHooks/useSessionExpired';
-import Spinner from '../../common/Spinner/index';
-import EditContest from './EditContest';
 import SuperuserContainer from '../SuperuserContainer';
+import EditContestContainer from './EditContestContainer';
 
+// SuperuserContainer will check whether the user is superuser or not and
+// if the user is admin only then the user will be allowed to see the
+// component
+const EditContestPage = () => (
+  <SuperuserContainer
+    component={<EditContestContainer />}
+  />
+);
 
-const EditcontestContainer = () => {
-  const { redirectOnSessionExpiredBeforeRender, isSessionExpired } = useSessionExpired();
-  const { contestId } = useParams();
-  const {
-    loading, error, data,
-  } = useQuery(GET_CONTEST_EDIT_DETAILS, {
-    variables: { code: contestId },
-  });
-
-  if (loading) return <Spinner />;
-  if (error) return <SomethingWentWrong message="An error has been encountered." />;
-  if (data.contestCode) {
-    const contestData = data.contestCode;
-    const { users } = data;
-    return (
-      <SuperuserContainer>
-        <EditContest contestData={contestData} users={users} />
-      </SuperuserContainer>
-    );
-  }
-  if (isSessionExpired(data.contestCode)) {
-    // since the component hasn't rendered or returned anything,
-    // we use redirectOnSessionExpiredBeforeRender function
-    return redirectOnSessionExpiredBeforeRender();
-  }
-  // case for the user not being admin or superuser
-  return <SomethingWentWrong message="An unexpected error has been encountered" />;
-};
-
-export default EditcontestContainer;
+export default EditContestPage;
