+++ src/Components/superuser/ratings/index.jsx
@@ -1,40 +1,14 @@
 import React from 'react';
-import { useQuery } from '@apollo/react-hooks';
-import { GET_ALL_CONTEST_DETAILS } from '../../../graphql/queries';
-import SomethingWentWrong from '../../common/SomethingWentWrong/index';
-import useSessionExpired from '../../../customHooks/useSessionExpired';
-import Spinner from '../../common/Spinner/index';
-import Ratings from './Ratings';
 import SuperuserContainer from '../SuperuserContainer';
+import RatingsContainer from './RatingsContainer';
 
+// SuperuserContainer will check whether the user is superuser or not and
+// if the user is admin only then the user will be allowed to see the
+// component
+const RatingsPage = () => (
+  <SuperuserContainer
+    component={<RatingsContainer />}
+  />
+);
 
-const RatingsContainer = () => {
-  const { redirectOnSessionExpiredBeforeRender, isSessionExpired } = useSessionExpired();
-  const {
-    loading, error, data,
-  } = useQuery(GET_ALL_CONTEST_DETAILS, {
-    variables: { limit: 150, skip: 0 },
-  });
-
-  if (loading) return <Spinner />;
-  if (error) return <SomethingWentWrong message="An error has been encountered." />;
-  if (data.allContests) {
-    const { contests } = data.allContests;
-    return (
-      <>
-        <SuperuserContainer>
-          <Ratings contests={contests} />
-        </SuperuserContainer>
-      </>
-    );
-  }
-  if (isSessionExpired(data.allContests)) {
-    // since the component hasn't rendered or returned anything,
-    // we use redirectOnSessionExpiredBeforeRender function
-    return redirectOnSessionExpiredBeforeRender();
-  }
-  // case for the user not being admin or superuser
-  return <SomethingWentWrong message="An unexpected error has been encountered" />;
-};
-
-export default RatingsContainer;
+export default RatingsPage;
